feat: allow to use a custom mnemonic for the 12 words seed phrase
