+++ lib/app/router/routs/add_seed/add_seed.dart
@@ -6,7 +6,7 @@ import 'package:app/feature/add_seed/add_existing_wallet/view/add_existing_walle
 import 'package:app/feature/add_seed/add_seed_enable_biometry/add_seed_enable_biometry_page.dart';
 import 'package:app/feature/add_seed/create_password/screens/create_seed_password/create_seed_password_screen.dart';
 import 'package:app/feature/add_seed/create_password/view/create_seed_password_page.dart';
-import 'package:app/feature/add_seed/enter_seed_phrase/enter_seed_phrase.dart';
+import 'package:app/feature/add_seed/enter_seed_phrase/enter_seed_phrase_widget.dart';
 import 'package:app/feature/add_seed/import_wallet/import_wallet_screen.dart';
 import 'package:app/v1/feature/add_seed/check_seed_phrase/check_seed_phrase.dart';
 import 'package:app/v1/feature/add_seed/create_seed/create_seed.dart';
@@ -58,7 +58,7 @@ GoRoute createSeedNoNamedRoute(GoRoute passwordRoute) {
 GoRoute enterSeedNoNamedRoute(GoRoute passwordRoute) {
   return GoRoute(
     path: AppRoute.enterSeed.path,
-    builder: (_, __) => const EnterSeedPhrase(),
+    builder: (_, __) => const EnterSeedPhraseWidget(),
     routes: [
       passwordRoute,
     ],
@@ -200,7 +200,7 @@ GoRoute get createSeedNamedProfileRoute {
 GoRoute get enterSeedNamedProfileRoute {
   return GoRoute(
     path: AppRoute.enterSeedNamed.path,
-    builder: (_, state) => const EnterSeedPhrase(),
+    builder: (_, state) => const EnterSeedPhraseWidget(),
     routes: [
       GoRoute(
         path: AppRoute.createSeedPassword.path,
