+++ lib/feature/add_seed/add_seed_enable_biometry/add_seed_enable_biometry.dart
@@ -1,0 +1,1 @@
+export 'add_seed_enable_biometry_page.dart';
