+++ lib/feature/add_seed/create_password/create_password.dart
@@ -1,2 +1,5 @@
 export 'cubit/cubit.dart';
+export 'model/model.dart';
+export 'screens/screens.dart';
 export 'view/view.dart';
+export 'widgets/widgets.dart';
