+++ lib/feature/add_seed/create_password/cubit/cubit.dart
@@ -1,2 +1,1 @@
-//GENERATED BARREL FILE
 export 'create_seed_password_cubit.dart';
