+++ lib/feature/add_seed/create_password/model/model.dart
@@ -1,0 +1,1 @@
+export 'password_status.dart';
