+++ lib/feature/add_seed/create_password/screens/create_seed_password/create_seed_password.dart
@@ -1,0 +1,3 @@
+export 'create_seed_password_screen.dart';
+export 'create_seed_password_screen_model.dart';
+export 'create_seed_password_screen_wm.dart';
