+++ lib/feature/add_seed/create_password/screens/screens.dart
@@ -1,0 +1,1 @@
+export 'create_seed_password/create_seed_password.dart';
