+++ lib/feature/add_seed/create_password/view/view.dart
@@ -1,3 +1,2 @@
-//GENERATED BARREL FILE
 export 'create_seed_password_page.dart';
 export 'create_seed_password_view.dart';
