+++ lib/feature/add_seed/create_password/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'password_info_section.dart';
