+++ lib/feature/add_seed/enter_seed_phrase/data/data.dart
@@ -1,0 +1,2 @@
+export 'input_data.dart';
+export 'tab_data.dart';
