+++ lib/feature/add_seed/enter_seed_phrase/widgets/widgets.dart
@@ -1,0 +1,2 @@
+export 'tabs.dart';
+export 'words.dart';
