+++ lib/feature/add_seed/import_wallet/import_wallet_screen.dart
@@ -2,10 +2,7 @@
 // ignore_for_file: invalid_use_of_protected_member
 
 import 'package:app/app/router/router.dart';
-import 'package:app/data/models/models.dart';
-import 'package:app/feature/add_seed/import_wallet/data/import_wallet_data.dart';
-import 'package:app/feature/add_seed/import_wallet/import_wallet_widget_model.dart';
-import 'package:app/feature/constants.dart';
+import 'package:app/feature/add_seed/add_seed.dart';
 import 'package:app/generated/generated.dart';
 import 'package:elementary/elementary.dart';
 import 'package:elementary_helper/elementary_helper.dart';
@@ -146,7 +143,7 @@ class ImportWalletScreen
                     children: [
                       StateNotifierBuilder(
                         listenableState: wm.seedPhraseFormat,
-                        builder: (_, seedPhraseFormat) => _SeedPhraseFormat(
+                        builder: (_, seedPhraseFormat) => SeedPhraseFormatView(
                           networkGroup: wm.networkGroup,
                           wordsCount: data?.selectedValue,
                           value: seedPhraseFormat,
@@ -166,80 +163,6 @@ class ImportWalletScreen
             ),
           );
         },
-      ),
-    );
-  }
-}
-
-class _SeedPhraseFormat extends StatelessWidget {
-  const _SeedPhraseFormat({
-    required this.networkGroup,
-    required this.wordsCount,
-    required this.value,
-    required this.onChanged,
-  });
-
-  final String networkGroup;
-  final int? wordsCount;
-  final SeedPhraseFormat? value;
-  final ValueChanged<SeedPhraseFormat>? onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = context.themeStyleV2;
-
-    // Seed phrase format selector is available only for
-    // TON and HMSTR network and when seed phrase length is 24
-    if (wordsCount != legacySeedPhraseLength ||
-        (networkGroup != 'ton' && networkGroup != 'hmstr_mainnet')) {
-      return const SizedBox.shrink();
-    }
-
-    return Padding(
-      padding: const EdgeInsets.only(
-        bottom: DimensSizeV2.d16,
-      ),
-      child: Container(
-        padding: const EdgeInsets.all(DimensSizeV2.d16),
-        decoration: BoxDecoration(
-          border: SquircleBoxBorder(
-            squircleRadius: DimensRadiusV2.radius12,
-            borderSide: BorderSide(color: theme.colors.border0),
-          ),
-        ),
-        child: SeparatedColumn(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          separatorSize: DimensSizeV2.d12,
-          children: [
-            Text(
-              LocaleKeys.seedPhraseFormatLabel.tr(),
-              style: theme.textStyles.labelSmall,
-            ),
-            SeparatedRow(
-              separatorSize: DimensSizeV2.d16,
-              children: [
-                CommonRadioInput(
-                  value: SeedPhraseFormat.standart,
-                  groupValue: value,
-                  onChanged: onChanged,
-                  child: Text(
-                    LocaleKeys.seedPhraseFormatStandart.tr(),
-                    style: theme.textStyles.labelSmall,
-                  ),
-                ),
-                CommonRadioInput(
-                  value: SeedPhraseFormat.bip39,
-                  groupValue: value,
-                  onChanged: onChanged,
-                  child: Text(
-                    LocaleKeys.seedPhraseFormatBIP39.tr(),
-                    style: theme.textStyles.labelSmall,
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
       ),
     );
   }
