+++ lib/feature/add_seed/view/view.dart
@@ -1,0 +1,1 @@
+export 'seed_phrase_format_view.dart';
